requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
awscli==1.35.19
boto3==1.35.53
botocore==1.35.53
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==5.0.7
django-cors-headers==4.4.0
django-request-logging==0.7.5
djangorestframework==3.15.2
docker==7.1.0
docutils==0.16
et-xmlfile==1.1.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
jaraco.classes==3.4.0
jaraco.context==6.0.1
Jinja2==3.1.4
jmespath==1.0.1
keras==3.4.1
keyring==8.7
keyrings.alt==5.0.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
more-itertools==10.5.0
mpmath==1.3.0
mysqlclient==2.2.4
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==11.0.0
protobuf==4.25.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3transfer==0.10.3
scipy==1.14.1
seaborn==0.13.2
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.27
ultralytics-thop==2.0.10
urllib3==2.2.2
voluptuous==0.15.2
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
yolo==0.3.1
